scrape: properly retry on 500 errors
The Art Blocks token API seems to sometimes respond with a 500 Internal
Server Error whose body is `{"message":"Token Not Found"}`, even for
tokens that definitely exist. We now properly retry in such cases.

You can detect this pretty reliably by bumping the concurrency to 256
and ingesting a large project.