db: try to roll back when releasing clients
If a connection is released to the pool while in an aborted-transaction
state, that state is not reset before the pool is provided to a new
client. This can poison future SQL operations for a batch of operations
(like `add-project-tokens all`). This patch changes `acqrel` to attempt
to roll back any active transaction. (When no transaction is in
progress, `ROLLBACK` is a no-op.)