scrape: use multiline regexes to extract data
In particular, descriptions can span multiple lines (e.g., project 11).
As written, trying to read the description fails in such cases,
basically because `"a\nb".match(/^a.*$/) == null`. We currently only
read the description if the artist name is empty, so we don't really hit
this issue, but soon we'll want to read the description in all cases.