util: add `combo` parser library
Licensing: I wrote this code in its entirety, and [originally published
it for SourceCred][1] under the terms of that project. I reuse it here
and provide it under the terms of this project. This code is forked from
its state at SourceCred commit fb669962a030; I was the sole author of
all changes to the file through that point.

[1]: https://github.com/sourcecred/sourcecred/pull/1816