img: make ingestion incremental
We now treat update the GCS listing cache as we populate targets, and
run ingestion in an infinite loop.

To test, create a fresh database and run `node . add-project 22` and
`node . add-token 22000000`. Then, start the `ingest-images` task. Watch
it process the added token and then go into a waiting loop. In another
shell, run `node . add-token 22000001`, then watch as the ingester picks
that up, too.