cli: add `follow-live-mint` command
At runtime, this has two states: "waiting for next token" and "gobbling
up tokens". While waiting for next token, it slowly polls the lowest
unfetched token for the mint until that token resolves, at which point
it switches into gobbling. While gobbling, it fetches bunches of tokens
more aggressively and in parallel; as soon as one of these 404s, it
switches back to waiting for the next token.

Tested by adding

```javascript
  if (Number(String(await require("fs").promises.readFile("/tmp/maxtoken"))) < tokenId) return null;
```

to `fetchTokenJsonText` in `fetchArtblocksToken.js` and manually poking
that file while the job is running to simulate an active mint.