scrape: block empty features during live mint
It turns out that not only are there a handful of Art Blocks projects
that genuinely have no features for any token, there are also a handful
of tokens across various projects that have no features even though all
their siblings do: the Art Blocks API is just returning the wrong data.
To account for this, we make two changes:

  - We maintain a whitelist of projects that are definitely allowed to
    have empty `features`: e.g., Elevated Deconstructions.
  - If we see token data without `features` for a project _not_ on this
    whitelist, then we allow the token to be added if backfilling the
    database (`add-project-tokens all`) but not during a live mint
    (since it could be a transient issue, and likely is).

To test, try running `follow-live-mint` on projects 7 and 193, on a new
database. Project 7 should proceed swimmingly, because Elevated
Deconstructions are expected to not have features. Project 193 should
fail upon reaching token #3, because that one is just missing features.