db: add `(token_contract, on_chain_token_id)` IDs
Our current `token_id` field is really an on-chain ID, which is not
unique across contract addresses, and is also stored in a data type that
is theoretically too small. We'll migrate data and usage over to these
new fields, preserving current semantics, and then add a new `token_id`
that is globally unique and migrate to that.

These two new fields will stay on `tokens`, but will eventually be
removed from `trait_members` once we can just use (the new) `token_id`.