db: fix project ID filtering for tokens with no traits
Before this patch, a `getTokenFeaturesAndTraits` call with a project ID
filter would also include matches for tokens with no traits. As a
result, `GET_LATEST_TOKENS` websocket messages with `lastTokenId: null`,
which are processed with a `{ projectId, minTokenId: 0, maxTokenId }`
filter, could receive responses from tokens of arbitrary earlier
projects. This patch fixes that.