replace ingest_opensea_events with ingest_opensea_collection
clarify that this method only gets a single collection, since we'll have
a new command that gets events across all AB projects.

also, simplify the CLI -- we no longer manually specify a start time or
scan window, we default to scanning from the beginning of time, and
always scanning in monthlong increments.