db: change `token_index` from `int8` to `int`
This is reasonable because it seems unrealistic that a single collection
of generative art would have more than 2 billion pieces. It's convenient
because it means that token indices will be returned from `pg` queries
as JS numbers instead of numeric strings. And it's safe because the
column is so far only used in input positions and never in the output of
a query, so no output types should change.