api: return project newids and Art Blocks indices
A concession. The legacy `ab-#` IDs were being scraped for Art Blocks
project index by the frontend. The frontend still actually needs that
for now, so let's give it to them in a more explicit form: a property
called `artblocksProjectIndex`. We also now provide the project newid.
The goal is to move off of the base `projectId` and its `ab-#` form.