ingest-images: bracket `listenClient` acquisition
This was originally unbracketed because the thing is meant to run
forever, but we still want to release the `listenClient` if we have to
unwind due to an exception. In such a case, the code now successfully
exits rather than hanging on `withDb`'s `pool.end`.

This patch is mostly whitespace; suggest reviewing with `-w`.