db: roll up historical migrations
A new `rollup.sql` script brings an empty database to the same state
that it would have if we applied the first 50 migrations in sequence.
This script is generated by applying the migrations and dumping the
database with `pg_dump(1)`. This is smarter than just concatenating the
SQL text of the migration scripts, because it only describes the final
state: e.g., if you create a table and then drop it, the table will not
appear at all in `rollup.sql`.

Applying this rollup to an empty database is substantially faster than
applying all migrations in sequence. On my machine, it brings
initialization time down from ~700 ms to ~200 ms. Since we initialize a
fresh database on every test case, this makes a big difference in total
test execution time: for me, it cuts it roughly in half (40s to 21s).

To regenerate `rollup.sql`, run `src/db/migrations/updateRollup.js`.
This should not change its semantics unless you also change the value of
`lastMigrationInRollupName` in `migrations/index.js`.

A test in `migrations/index.test.js` verifies that the rollup state and
the from-scratch state are the same.