opensea: retrieve floor-k tokens with current-owner check
We've discovered that OpenSea's APIs incorrectly omit some cancellation
events when you request them for a collection (by either collection slug
or token contract), but return them when you ask for cancellations on a
specific token. So, this patch adds a simple little query to get, for
each collection, the `k` tokens with lowest-priced open asks by their
current holders, so that we can interrogate OpenSea and see if those
asks are really still valid.

It's in a new `opensea/hacks.js` module because I haven't written tests
for it yet and don't want to pollute the well-crafted coverage of the
main module.

Paired with @decentralion.