followChainTokens: listen for deferred transfers
We now send a pubsub message when we defer an ERC-721 transfer event.
Thus, when listening for new deferrals to read from the Art Blocks API,
we only need to wake when we receive such an event. We still wake every
minute unconditionally to handle the case where our first query to the
Art Blocks API is before it's ready.