sortAsciinumeric: naturally order small English numbers
The impetus of this change is the "Bouquet" feature set on the "Chimera"
project, which is currently sorted thus:

    Eight flowers
    Five flowers
    Four flowers
    Nine flowers
    None
    One flower
    ...

This is clearly sad. After this change, it's instead sorted thus:

    None
    One flower
    Two flowers
    ...

Of course, this also works for features sets that just have the numbers
without any adjoined noun, like `["None", "One", "Two", "Three"]` from
Chimera's "Books" feature.

As a nice bonus, feature sets that include a bunch of non-numeric
categories plus "None" now see "None" at the front instead of
alphabetized with the "N"s, which seems nice to me.

The ordering is somewhat worsened for semantically categorical sets that
just happen to contain some numbers, like the "Center" trait from the
project called "8", which is now sorted thus:

    Zero & One
    Two Moons
    Eight Ball
    Blank
    Cherries
    Choose Luck or Fate
    ...

It seems slightly weird to have "Two Moons" and "Eight Ball" before
"Cherries", but hey, them's the breaks. I don't make the rules; I just
make the rules.