cnfs: rewrite canonicalization logic without stringification
We now compare arrays lexicographically using `Cmp.array` instead of
stringifying them and relying on the string representation having the
right natural order in this particular case. This feels more principled
to me and also requires fewer intermediate allocations and data
structures (no sets, no marshalling, just in-place array ops). The whole
thing could be in-place, too, except that `canonicalForm` itself is not
meant to mutate its arguments, so we deep-copy explicitly.