Opensea: skip asks and cancellations w/ no listing time
This addresses a new data corruption issue we're seeing where
older/historical events don't have a listing time. We think this is
associated with OpenSea recently hackily killing all old listings.

We depend on the listing time as a sort of knockoff identifier for asks,
hence needing them on the cancellations too. I considered attempting to
refactor this code, but since the issue only affects old asks (which are
all cancelled anyway due to the aforementioned hack) I think just
dropping them is fine. Sales with null listing time are OK.