chain: fix parent hash check for first-fetched block
(Well, I suspected that it would eventually bite me to use field names
`blockHash` and `blockNumber` instead of `hash` and `number`.) This was
causing the local head parent check to be a no-op, because the expected
hash would be `undefined`. If the chain reorgs between `unrollReorgs`
and `addNewHeaders`, this can fail to detect the reorg. Due to the
foreign key constraint on `parent_hash`, we would hard-fail:

```
[2022-05-31T19:33:49.564Z DEBUG eth/chain::addNewHeaders] expecting parent hash undefined based on local head
[2022-05-31T19:33:49.564Z DEBUG eth/chain::addNewHeaders] adding block #14880531 with hash 0xa9c2929832f413d985ce6833294e37d1bb0d952cbab2cdd6aeeba6947f84ed1c, parent 0x5d5fb5964bea2aad4881eb9b3906919bd45c419b56a0a3774cc52110abe069d9
[2022-05-31T19:33:49.568Z ERROR db/util] withPool callback failed: error: insert or update on table "eth_blocks" violates foreign key constraint "eth_blocks_parent_hash_fkey": Error
```

Instead, we should now hit the "retry" case and go back to unrolling
reorgs.