db: remove broken and unneeded parent hash constraint
This is broken because it checks against `'0x00...00'::bytea`, which is
actually `'\x30783030...3030'::bytea` instead of `'\x00...00'::bytea`
due to the leading `0x` instead of `\x`. (Could have caught this by
explicitly casting the literal  in the `CHECK` constraintto `bytes32`.)

It's unnecessary because the intended purpose was just to ensure that
the genesis block doesn't get inserted with a zero `parent_hash`
(instead of `NULL`), but such an error would already be caught due to
(a) the `(parent_hash IS NULL) = (block_number = 0)` check constraint,
and (b) the `parent_hash REFERENCES eth_blocks(block_hash)` foreign key
constraint, since no block has the zero hash.