eth: `queryFilter` by hash for recent blocks
If we use `contract.queryFilter(filter, minBlock, maxBlock)` and our
query hits a remote node that is lagging behind, it may answer our query
for non-canonical blocks in the `minBlock..=maxBlock` range. If this
causes non-canonical events to be returned, we'll detect that, because
the block hashes in the events will not match those in our database.
However, if this causes canonical events to *not* be returned because
they're not in the non-canonical blocks, then we'll just get an empty
list back and have no way to tell that something went wrong.

To work around this, `safeQueryFilter` partitions such a `queryFilter`
call into two families of sub-calls: a single range-based call over
blocks that are old enough to be considered finalized, and zero or more
hash-based calls for more recent blocks queried directly by block hash.
Frequently, only one of these families will actually have calls: if the
`queryFilter` is entirely historical then only the range-based call will
be present, and if it's entirely recent then only the point calls will
be present.