orderbook: factor out `BID_PLACED` message construction
...with intention to reuse this logic when previously inactive bids are
reactivated. This cleans up `addBid` fairly nicely, too, and because we
can get all the data for the notification in this one query while
simultaneously dropping the `getSlug`/etc. cruft, it should decrease the
database load at the same time.

Existing tests cover this nicely, for all four scope types.